Add theming, internationalization, bloc state management (#37)
These changes provide theming, internationalization and
state management changes. The internationalization is done
with i10n (the default of Flutter), the theming is provided
both to the admin by changing theming files and to simple
users by using their settings page. Th BLoC pattern is used
to keep track of the state of the application.